wip: changing client-server communication

+++ client/client.go
@@ -5,8 +5,8 @@ import (
 )
 
 type Vector struct {
-	X int `json:"x"`
-	Y int `json:"y"`
+	X uint16 `json:"x"`
+	Y uint16 `json:"y"`
 }
 
 type Player struct {
